feat(auth): add authorization command
Incorporate a modified version of the Mutt OAuth2 token script, with the
`localhostauth` method hard-coded. The email may be provided via the
command line, else it is prompted from the user.

Make a `.thallo` directory in the home directory and keep the access
token in there. Use a `utils` module to abstract access to it.