feat: replace day with info
The new `info` command has a little more versatility than `day` did, in
that it can be used to query for specific events via the `--index` or
`--name` flags. Specific events are printed with a higher level of
detail, including the body of the event and a list of attendees.